Update netclient_linux_upg.sh
added 
- checks for current version
- user input for accepting update
- remove backup if exists

+++ netclient_linux_upg.sh
@@ -2,24 +2,40 @@
 
 # upgrade netclient on linux clients
 
+# Find the current netclient version
+netclient_tag=$(curl -sL https://api.github.com/repos/gravitl/netmaker/releases/latest | jq -r ".tag_name")
+echo -----------------------------------------
+echo current version of netclient is $netclient_tag
+echo -----------------------------------------
+read -p  "Do you wish to install this version?" -n 1 -r
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+    exit 1
+fi
+
 # stop service
 systemctl stop netclient.service
 
 # move to netclient directory
 cd /etc/netclient
 
-# back up old binary
+# if there's an old backup, remove it
+FILE=/etc/netclient.bak
+if [[ -f "$FILE" ]]; then
+  rm $FILE
+  exit 0
+
+# back up current binary
 mv netclient netclient.bak
 
-# remember to update to correct version number in next line
-curl -L https://github.com/gravitl/netmaker/releases/download/v0.15.0/netclient --output /etc/netclient/netclient
+# download the new version into correct directory
+curl -L https://github.com/gravitl/netmaker/releases/download/$netclient_tag/netclient --output /etc/netclient/netclient
 
 # make new binary executable
 chmod a+x netclient
 
 # restart service
 systemctl start netclient.service
-
 
 # get new config
 netclient pull
